working conhost:posh and nui:core
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=097383a88b33dcf3f9683dcfd1484d57a3ddd5b1

+++ client/shared/Hooking.h
@@ -115,7 +115,7 @@ inline void return_function(AddressType address, uint16_t stackSize = 0)
 	else
 	{
 		put<uint8_t>(address, 0xC2);
-		put<uint16_t>(address + 1, stackSize);
+		put<uint16_t>((uintptr_t)address + 1, stackSize);
 	}
 }
